Small additions

+++ tempodb/tempodb.go
@@ -262,7 +262,7 @@ func (rw *readerWriter) Find(ctx context.Context, tenantID string, id encoding.I
 	foundBytes, err := rw.pool.RunJobs(copiedBlocklist, func(payload interface{}) ([]byte, error) {
 		meta := payload.(*encoding.BlockMeta)
 
-		shardKey := bloom.ShardKeyForTraceID(id)
+		shardKey := uint64(bloom.ShardKeyForTraceID(id))
 		level.Debug(logger).Log("msg", "fetching bloom", "shardKey", shardKey)
 		bloomBytes, err := rw.r.Bloom(meta.BlockID, tenantID, shardKey)
 		if err != nil {
